'scale.pl' now supports setting a module-name instead of 'LOGO'.
There was no issue with libraries created with 'make_lib.sh', as
that script sets a module-name based on the input-filename.

But 'scale.pl' left 'LOGO' untouched. This becomes an issue if
multiple footprints named 'LOGO' are inserted into a board and
a footprint-archive is created. Same-name modules got overwritten.